Refactor out traffic library functions
